fix the mac list problem in the previous commit
That entry is not a random entry, it caused by copy a struct contains
a list head. As the address of head in the new copy is changed.
But the double link list still use the old address in it, so it would
failed to find it is head of list.